requirements.txt
python-dotenv
ics